requirements.txt
aiogram==3.15.0
SQLAlchemy==2.0.32
aiosqlite==0.20.0
python-dotenv==1.0.1
pytz==2024.1
APScheduler==3.10.4
greenlet==3.2.4
Flask==3.0.3

